+++ tests/tests_cycle_decomposition/test_generic_methods.py
@@ -9,6 +9,7 @@ from tests.test_factory import (
     validate_inverse,
     validate_is_even,
     validate_support,
+    validate_cycle_type,
     validate_equivalent,
     validate_cycle_notation,
     validate_is_derangement,
@@ -23,6 +24,7 @@ from tests.tests_cycle_decomposition.test_cases import (
     TEST_INVERSE,
     TEST_IS_EVEN,
     TEST_SUPPORT,
+    TEST_CYCLE_TYPE,
     TEST_EQUIVALENT,
     TEST_CYCLE_NOTATION,
     TEST_IS_DERANGEMENT,
@@ -48,6 +50,16 @@ def test_cycle_decomposition(cycle_decomposition, expected_value) -> None:
 def test_cycle_notation(cycle_decomposition, expected_value) -> None:
     """Tests for the method `cycle_notation()`."""
     validate_cycle_notation(item=cycle_decomposition, expected_value=expected_value)
+
+
+@pytest.mark.parametrize(
+    argnames="cycle_decomposition, expected_value",
+    argvalues=TEST_CYCLE_TYPE,
+    ids=[f"{c}.cycle_type()={s}" for c, s in TEST_CYCLE_TYPE],
+)
+def test_cycle_type(cycle_decomposition, expected_value) -> None:
+    """Tests for the method `cycle_type()`."""
+    validate_cycle_type(item=cycle_decomposition, expected_value=expected_value)
 
 
 @pytest.mark.parametrize(